data mod moved to service/query_serivce
