--- src/data/mod.rs
@@ -1,5 +1,0 @@
-mod table_builder;
-mod column_builder;
-
-pub use table_builder::*;
-pub use column_builder::*;
