+++ src/service/mod.rs
@@ -1,0 +1,3 @@
+mod query_service;
+
+pub use query_service::QueryService;
